backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
celery==5.3.3
redis==5.0.0 
hiredis==2.2.3 
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
